+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Now you can set a suffix-icon near the default arrow icon
 * Added `text-align` with default value `center` to multiple `Select`'s collapser dropdown
 * Created validation functionality. Additional info you can find in storybook validation notes.
 * Created `NotEmptyRule` `LengthInRange` `PassRegexpRule` rules for validation.
-* Added validation for `st-input`
+* Added validation for `Input`
+* Added `optionClass` prop to `Select`
 
 ### Maintenance
 * Updated `Notification` component. Do not show browser notifications when window is active